[SYCL][ABI Break] Remove ext::oneapi::reduction
This PR also move the implementation from the extension header to the
regular sycl/reduction.hpp.

There might be some simplifications in the implementation enabled by
that but they are left for a future PR.
