+++ sycl/test/abi/sycl_symbols_linux.dump
@@ -3662,10 +3662,6 @@ _ZN4sycl3_V13ext6oneapi10level_zero12make_programERKNS0_7contextEm
 _ZN4sycl3_V13ext6oneapi10level_zero13make_platformEm
 _ZN4sycl3_V13ext6oneapi15filter_selectorC1ERKNSt7__cxx1112basic_stringIcSt11char_traitsIcESaIcEEE
 _ZN4sycl3_V13ext6oneapi15filter_selectorC2ERKNSt7__cxx1112basic_stringIcSt11char_traitsIcESaIcEEE
-_ZN4sycl3_V13ext6oneapi6detail16reduGetMaxWGSizeESt10shared_ptrINS0_6detail10queue_implEEm
-_ZN4sycl3_V13ext6oneapi6detail17reduComputeWGSizeEmmRm
-_ZN4sycl3_V13ext6oneapi6detail22reduGetPreferredWGSizeERSt10shared_ptrINS0_6detail10queue_implEEm
-_ZN4sycl3_V13ext6oneapi6detail33reduGetMaxNumConcurrentWorkGroupsESt10shared_ptrINS0_6detail10queue_implEE
 _ZN4sycl3_V14freeEPvRKNS0_5queueERKNS0_6detail13code_locationE
 _ZN4sycl3_V14freeEPvRKNS0_7contextERKNS0_6detail13code_locationE
 _ZN4sycl3_V15event13get_wait_listEv
@@ -3712,8 +3708,6 @@ _ZN4sycl3_V15queueC2ERKNS0_7contextERKNS0_6deviceERKNS0_13property_listE
 _ZN4sycl3_V15queueC2ERKNS0_7contextERKNS0_6deviceERKSt8functionIFvNS0_14exception_listEEERKNS0_13property_listE
 _ZN4sycl3_V16ONEAPI15filter_selectorC1ERKNSt7__cxx1112basic_stringIcSt11char_traitsIcESaIcEEE
 _ZN4sycl3_V16ONEAPI15filter_selectorC2ERKNSt7__cxx1112basic_stringIcSt11char_traitsIcESaIcEEE
-_ZN4sycl3_V16ONEAPI6detail16reduGetMaxWGSizeESt10shared_ptrINS0_6detail10queue_implEEm
-_ZN4sycl3_V16ONEAPI6detail17reduComputeWGSizeEmmRm
 _ZN4sycl3_V16detail10build_implERKNS0_13kernel_bundleILNS0_12bundle_stateE0EEERKSt6vectorINS0_6deviceESaIS8_EERKNS0_13property_listE
 _ZN4sycl3_V16detail10image_impl10getDevicesESt10shared_ptrINS1_12context_implEE
 _ZN4sycl3_V16detail10image_impl11allocateMemESt10shared_ptrINS1_12context_implEEbPvRP9_pi_event
@@ -3824,9 +3818,11 @@ _ZN4sycl3_V16detail15getOrWaitEventsESt6vectorINS0_5eventESaIS3_EESt10shared_ptr
 _ZN4sycl3_V16detail16AccessorImplHost6resizeEm
 _ZN4sycl3_V16detail16AccessorImplHostD1Ev
 _ZN4sycl3_V16detail16AccessorImplHostD2Ev
+_ZN4sycl3_V16detail16reduGetMaxWGSizeESt10shared_ptrINS1_10queue_implEEm
 _ZN4sycl3_V16detail17HostProfilingInfo3endEv
 _ZN4sycl3_V16detail17HostProfilingInfo5startEv
 _ZN4sycl3_V16detail17device_global_map3addEPKvPKc
+_ZN4sycl3_V16detail17reduComputeWGSizeEmmRm
 _ZN4sycl3_V16detail18convertChannelTypeE22_pi_image_channel_type
 _ZN4sycl3_V16detail18convertChannelTypeENS0_18image_channel_typeE
 _ZN4sycl3_V16detail18get_kernel_id_implENSt7__cxx1112basic_stringIcSt11char_traitsIcESaIcEEE
@@ -3847,6 +3843,7 @@ _ZN4sycl3_V16detail22get_kernel_bundle_implERKNS0_7contextERKSt6vectorINS0_6devi
 _ZN4sycl3_V16detail22get_kernel_bundle_implERKNS0_7contextERKSt6vectorINS0_6deviceESaIS6_EERKS5_INS0_9kernel_idESaISB_EENS0_12bundle_stateE
 _ZN4sycl3_V16detail22has_kernel_bundle_implERKNS0_7contextERKSt6vectorINS0_6deviceESaIS6_EENS0_12bundle_stateE
 _ZN4sycl3_V16detail22has_kernel_bundle_implERKNS0_7contextERKSt6vectorINS0_6deviceESaIS6_EERKS5_INS0_9kernel_idESaISB_EENS0_12bundle_stateE
+_ZN4sycl3_V16detail22reduGetPreferredWGSizeERSt10shared_ptrINS1_10queue_implEEm
 _ZN4sycl3_V16detail23constructorNotificationEPvS2_NS0_6access6targetENS3_4modeERKNS1_13code_locationE
 _ZN4sycl3_V16detail23getESIMDDeviceInterfaceEv
 _ZN4sycl3_V16detail24find_device_intersectionERKSt6vectorINS0_13kernel_bundleILNS0_12bundle_stateE1EEESaIS5_EE
@@ -3860,6 +3857,7 @@ _ZN4sycl3_V16detail2pi9getPluginILNS0_7backendE1EEERKNS1_6pluginEv
 _ZN4sycl3_V16detail2pi9getPluginILNS0_7backendE2EEERKNS1_6pluginEv
 _ZN4sycl3_V16detail2pi9getPluginILNS0_7backendE3EEERKNS1_6pluginEv
 _ZN4sycl3_V16detail2pi9getPluginILNS0_7backendE5EEERKNS1_6pluginEv
+_ZN4sycl3_V16detail33reduGetMaxNumConcurrentWorkGroupsESt10shared_ptrINS1_10queue_implEE
 _ZN4sycl3_V16detail36get_empty_interop_kernel_bundle_implERKNS0_7contextERKSt6vectorINS0_6deviceESaIS6_EE
 _ZN4sycl3_V16detail6OSUtil10getDirNameB5cxx11EPKc
 _ZN4sycl3_V16detail6OSUtil11alignedFreeEPv
